test(hpos-hal): feature-gate `sudo` tests, test without sandbox, add required programs
enabling `sudo` in the build process is (a) not a feature and would
require unsupported hacks in the nix build process and also risk test
machine integrity.

subsequently we can redesign the tests and rely on test fixtures and a
root prefix to read from a fixture root instead of the real one.