fix(host-agent): add server_name CLI flag and fallback to unique machine-id
this fixes an issue where the server_name is not unique in the context
of a remote server that is used by multiple host-agents simultaneously.

in that situation the outgoing leafnode connections preempt each other,
causing connection churn.

previously i had falsely assumed that leaving it empty would case NATS
use an auto-generated unique server_name.