feat(holo-agent nixos integration test): use 5 hosts
this extends the test scenario to 5 hosts to increase the chances to
surface any connectivity issues that are related to configuration.

it would have been possible to reduce repetitions here, however i'm not
sure where we're going with this test subsequently and it might be
premature optimization. e.g., maybe we want to give some hosts special
behavior over others and thus require special casing again.