fix(host-agent/leaf-server): don't drop the NATS server prematurely, actually terminate server
the just previously added `.kill_on_drop(true)` call had immediate
effect as the Command instance was dropped right away. storing the
`Child` is not enough to prevent that.

by storing the Command in the LeafServer as well and returning LeafServer instance to the caller,
the caller can ensure to keep the NATS instance around as long as
desired.

this fixes the integration test.

i tested the drop behavior manually by watching the processes while
starting/stopping the host-agent.

TODO: write a programmatic rust test to verify the NATS process is
properly terminated and cleaned up.