various nixos module and container improvements
* feat(devshells/rust): add holochain to packages

holochain and related can be helpful when working on the holochain based
workloads

* feat(nixos host-agent module): individual pkg, nix builds, log settings

* feat(holo-nats-server nixos module): host config option, fallback TLS

falling back to internal TLS certificates is helpful for the testing
environments so that TLS can be used there more conveniently

* nix/holochain modules): feature override args; refactor config options

these variables are required for holochain workloads in real-world
scenarios. for now we support these via container overrides.

the refactor makes it possible to set various networking options via the
module options granularly without having to override the entire
conductor config.

* fix(rust-workspace): openssl in cargo-nextest as well

this is surprisingly needed. we can debug this subsequently.